requirements.txt
streamlit
google-generativeai
python-dotenv
pandas 
numpy
flask
ipython
Pillow